backend/pyproject.toml
[tool.poetry]
name = "website-blockchain-backend"
version = "0.1.0"
description = ""
authors = ["Ravi Deevan <ravideevan@duck.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pyskein = "^1.0"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
